:elf: Use elf entity store for prices
