+++ shared/stores/prices/price.model.ts
@@ -1,0 +1,8 @@
+export interface Price {
+  id: number;
+  priceKey: string;
+  displayName: string;
+  forOffers: boolean;
+  forWishes: boolean;
+  withAmount: number;
+}
